+++ src/du/du.rs
@@ -1,5 +1,5 @@
 #![crate_name = "du"]
-#![feature(collections, core, io, path, rustc_private, std_misc, unicode)]
+#![feature(collections, core, old_io, old_path, rustc_private, std_misc, unicode)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -90,7 +90,7 @@ fn du(path: &Path, mut my_stat: Stat,
     stats
 }
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let program = args[0].as_slice();
     let opts = [
         // In task
