+++ src/sort/sort.rs
@@ -1,5 +1,5 @@
 #![crate_name = "sort"]
-#![feature(collections, core, io, path, rustc_private, unicode)]
+#![feature(collections, core, old_io, old_path, rustc_private, unicode)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -29,7 +29,7 @@ static VERSION:  &'static str = "0.0.1";
 static DECIMAL_PT: char = '.';
 static THOUSANDS_SEP: char = ',';
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let program = args[0].as_slice();
     let opts = [
         getopts::optflag("n", "numeric-sort", "compare according to string numerical value"),
