+++ src/tr/tr.rs
@@ -1,5 +1,5 @@
 #![crate_name = "tr"]
-#![feature(collections, core, io, rustc_private)]
+#![feature(collections, core, old_io, rustc_private)]
 
 /*
  * This file is part of the uutils coreutils package.
@@ -161,7 +161,7 @@ fn usage(opts: &[OptGroup]) {
     print(getopts::usage("Translate or delete characters.", opts).as_slice());
 }
 
-pub fn uumain(args: Vec<String>) -> isize {
+pub fn uumain(args: Vec<String>) -> i32 {
     let opts = [
         getopts::optflag("c", "complement", "use the complement of SET1"),
         getopts::optflag("C", "", "same as -c"),
