+++ src/true/true.rs
@@ -9,6 +9,6 @@
  * file that was distributed with this source code.
  */
 
-pub fn uumain(_: Vec<String>) -> isize {
+pub fn uumain(_: Vec<String>) -> i32 {
     0
 }
